Revise build/buildvc.bat and build/buildvc_x64.bat - Based on a report by Bert Mariani

+++ build/buildvc.bat
@@ -57,20 +57,20 @@ cd ..\..
 
 echo Building RingQt (Core)...
 cd ringqt
-del /q debug\*.* >nul
-del /q release\*.* >nul
+if exist debug\*.* del /q debug\*.* >nul
+if exist release\*.* del /q release\*.* >nul
 @call gencode_core >nul
 @call buildvc_core >nul
 
 echo Building RingQt (LightGUILib)...
-del /q debug\*.* >nul
-del /q release\*.* >nul
+if exist debug\*.* del /q debug\*.* >nul
+if exist release\*.* del /q release\*.* >nul
 @call gencode_light >nul
 @call buildvc_light >nul
 
 echo Building RingQt (GUILib)...
-del /q debug\*.* >nul
-del /q release\*.* >nul
+if exist debug\*.* del /q debug\*.* >nul
+if exist release\*.* del /q release\*.* >nul
 @call gencode_nobluetooth >nul
 @call buildvc_nobluetooth >nul
 cd binupdate
